Added warning for removing contact button

+++ code/Hyre_Defyer_Project_SE2/src/edu/westga/cs3211/hyre_defyer_project/view/DirectMessageView.java
@@ -9,10 +9,13 @@ import edu.westga.cs3211.hyre_defyer_project.view_model.FreelancerPostPageViewMo
 import javafx.collections.FXCollections;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.Label;
 import javafx.scene.control.ListView;
 import javafx.scene.control.TextArea;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.image.ImageView;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.AnchorPane;
@@ -103,8 +106,14 @@ public class DirectMessageView {
     
     @FXML
     void handleRemoveContactClick(ActionEvent event) {
-    	this.directMessageHandler.deleteChat(SignInViewModel.getCurrentUser(), this.contactListView.getSelectionModel().getSelectedItem());
-		this.updateContactList();
+    	Alert alert = new Alert(AlertType.CONFIRMATION);
+		alert.setHeaderText("Are you sure you want to remove this contact?");
+		alert.setContentText("Click OK to remove the contact, or Cancel to keep the contact.");
+		alert.showAndWait();
+		if (alert.getResult() == ButtonType.OK) {
+			this.directMessageHandler.deleteChat(SignInViewModel.getCurrentUser(), this.contactListView.getSelectionModel().getSelectedItem());
+			this.updateContactList();
+		}
 	}
 
     @FXML
